Make SPI work with 64 bit counters
Fix GET DIAGNOSTICS
Remove the earlier introduced SPI_processed64 variable

This includes the following upstream patches:
https://github.com/greenplum-db/gpdb/commit/23a27b039d94ba359286694831eafe03cd970eef
https://github.com/greenplum-db/gpdb/commit/f3f3aae4b7841f4dc51129691a7404a03eb55449
https://github.com/greenplum-db/gpdb/commit/ab737f6ba9fc0a26d32a95b115d5cd0e24a63191

THe https://github.com/greenplum-db/gpdb/commit/74a379b984d4df91acec2436a16c51caee3526af
is not yet included, because repalloc_huge() is not yet backported.

+++ src/backend/access/appendonly/aosegfiles.c
@@ -50,6 +50,7 @@ static void UpdateFileSegInfo_internal(Relation parentrel,
 						   int64 eof,
 						   int64 eof_uncompressed,
 						   int64 tuples_added,
+/* FIXME: diagnostics: int64 or uint64 */
 						   int64 varblocks_added,
 						   int64 modcount_added,
 						   FileSegInfoState newState);
@@ -232,6 +233,7 @@ GetFileSegInfo(Relation parentrel, Snapshot appendOnlyMetaDataSnapshot, int segn
 						fsinfo->eof, RelationGetRelationName(parentrel))));
 
 	/* get the tupcount */
+/* FIXME: diagnostics: int64 or uint64 */
 	fsinfo->total_tupcount = DatumGetInt64(
 										   fastgetattr(fstuple, Anum_pg_aoseg_tupcount, pg_aoseg_dsc, &isNull));
 
@@ -643,8 +645,11 @@ UpdateFileSegInfo(Relation parentrel,
 				  int segno,
 				  int64 eof,
 				  int64 eof_uncompressed,
+/* FIXME: diagnostics: int64 or uint64 */
 				  int64 tuples_added,
+/* FIXME: diagnostics: int64 or uint64 */
 				  int64 varblocks_added,
+/* FIXME: diagnostics: int64 or uint64 */
 				  int64 modcount_added,
 				  FileSegInfoState newState)
 {
@@ -677,8 +682,11 @@ UpdateFileSegInfo_internal(Relation parentrel,
 						   int segno,
 						   int64 eof,
 						   int64 eof_uncompressed,
+/* FIXME: diagnostics: int64 or uint64 */
 						   int64 tuples_added,
+/* FIXME: diagnostics: int64 or uint64 */
 						   int64 varblocks_added,
+/* FIXME: diagnostics: int64 or uint64 */
 						   int64 modcount_added,
 						   FileSegInfoState newState)
 {
@@ -953,6 +961,7 @@ GetSegFilesTotals(Relation parentrel, Snapshot appendOnlyMetaDataSnapshot)
  * Get the total bytes for a specific AO table from the pg_aoseg table on this local segdb.
  * (A version of GetSegFilesTotals that just gets the total bytes.)
  */
+/* FIXME: diagnostics: int64 or uint64 */
 int64
 GetAOTotalBytes(Relation parentrel, Snapshot appendOnlyMetaDataSnapshot)
 {
@@ -1165,8 +1174,9 @@ gp_update_aorow_master_stats_internal(Relation parentrel, Snapshot appendOnlyMet
 	Relation	aosegrel;
 	bool		connected = false;
 	char		aoseg_relname[NAMEDATALEN];
-	int			proc;
+	int			proc;	/* 32 bit, only holds number of segments */
 	int			ret;
+/* FIXME: diagnostics: int64 or uint64 */
 	int64		total_count = 0;
 	MemoryContext oldcontext = CurrentMemoryContext;
 
@@ -1200,7 +1210,7 @@ gp_update_aorow_master_stats_internal(Relation parentrel, Snapshot appendOnlyMet
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, false, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 
 		if (ret > 0 && SPI_tuptable != NULL)
@@ -1217,6 +1227,7 @@ gp_update_aorow_master_stats_internal(Relation parentrel, Snapshot appendOnlyMet
 				HeapTuple	tuple = tuptable->vals[i];
 				FileSegInfo *fsinfo = NULL;
 				int			qe_segno;
+/* FIXME: diagnostics: int64 or uint64 */
 				int64		qe_tupcount;
 				char	   *val_segno;
 				char	   *val_tupcount;
@@ -1454,6 +1465,7 @@ gp_aoseg_name(PG_FUNCTION_ARGS)
 
 		values[0] = Int32GetDatum(aoSegfile->segno);
 		values[1] = Int64GetDatum(aoSegfile->eof);
+/* FIXME: diagnostics: int64 or uint64 */
 		values[2] = Int64GetDatum(aoSegfile->total_tupcount);
 		values[3] = Int64GetDatum(aoSegfile->varblockcount);
 		values[4] = Int64GetDatum(aoSegfile->eof_uncompressed);
@@ -1554,8 +1566,10 @@ gp_update_ao_master_stats_oid(PG_FUNCTION_ARGS)
 
 typedef struct
 {
-	int			index;
-	int			rows;
+	uint64 index;
+	/* there is a chance the count will return more than 2^32 rows
+	 * plus SPI_processed is 64 bit anyway */
+	uint64 rows;
 } QueryInfo;
 
 
@@ -1982,7 +1996,7 @@ aorow_compression_ratio_internal(Relation parentrel)
 	Relation	aosegrel;
 	bool		connected = false;
 	char		aoseg_relname[NAMEDATALEN];
-	int			proc;
+	int			proc;	/* 32 bit, only holds number of segments */
 	int			ret;
 	float8		compress_ratio = -1;	/* the default, meaning "not
 										 * available" */
@@ -2026,7 +2040,7 @@ aorow_compression_ratio_internal(Relation parentrel)
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, false, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 		if (ret > 0 && SPI_tuptable != NULL)
 		{
